chore: refactor validators
